// src/api/generated/model/createRewardPhotoResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { RewardImageUploadDto } from './rewardImageUploadDto';

export interface CreateRewardPhotoResponseDto {
  /** ID созданной записи фото */
  photoId: string;
  upload: RewardImageUploadDto;
}